requirements.txt
# Basic build tools
setuptools>=65.5.0
wheel>=0.40.0

# Web framework and related - specify compatible versions
flask==2.3.3
werkzeug==2.3.7  # This version is compatible with Flask 2.3.3 and Flask-Session
gunicorn==21.2.0
flask-session==0.5.0
flask-cors==4.0.0

# Database
mysql-connector-python==8.1.0

# Data processing - use versions compatible with Python 3.12
numpy==2.0.2
pandas==2.2.2

# API and authentication
openai==1.72.0
python-dotenv==1.0.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0

# Development tools
pytest==7.4.0
black==23.7.0


# Add Google API dependencies
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1